Fix FileExplorerAndEditor AiEditor not resizing with viewport when viewport shrinks
